lesson5-4/pyproject.toml
[tool.poetry]
name = "exercise"
version = "0.1.0"
description = "Exercise"
authors = ["Hexlet"]

[tool.poetry.dependencies]
python = "^3.6"
hexlet-immutable-fs-trees = "*"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
